handle deletion in PromptHelper ausgelagert
